LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-transport-stats.https-expected.txt


FAIL DTLS statistics on transport-stats after setLocalDescription assert_equals: expected (string) "unknown" but got (undefined) undefined
FAIL ICE statistics on transport-stats after setLocalDescription assert_equals: expected (string) "controlling" but got (undefined) undefined
